feat(QW-05): add _MenuItemFormSheet for MenuScreen

+++ qwaiter_staff/lib/features/restaurant/menu_service.dart
@@ -90,7 +90,7 @@ class MenuService {
     String categoryID,
     String name,
     String description,
-    int price,
+    double price,
   ) async {
     try {
       await _dio.post(
@@ -125,7 +125,7 @@ class MenuService {
     String categoryID,
     String? name,
     String? description,
-    int? price,
+    double? price,
   ) async {
     try {
       final Map<String, dynamic> data = {
